+++ NEWS.md
@@ -1,5 +1,7 @@
 # NEWS
 ## 0.1.5.xxx
+- improve numeric to string conversion when using objects ids as names/dimnames (e.g. avoid 1e+05, use 100000 instead)
+
 - create popsRename function to simplify population renaming in `IFC_data` object
 
 - fix bug with popsCompute for type "C" (combined) population computation when population definition contains repeated occurrence of same population name